+++ lib/features/artists/artists_page.dart
